requirements.txt
PyQt5
spotipy
pytubefix
python-dotenv